metadata
- Pages now have unique metadata
- OpenGraph image added
- Favicon replaced
