+++ app/layout.tsx
@@ -1,10 +1,4 @@
-import type { Metadata } from "next";
 import "./globals.css";
-
-export const metadata: Metadata = {
-  title: "Hello",
-  description: "",
-};
 
 export default function RootLayout({
   children,
